+++ src/main/java/phanes/replay/user/strategy/SocialLoginStrategy.java
@@ -1,7 +1,7 @@
 package phanes.replay.user.strategy;
 
 import phanes.replay.user.domain.enums.SocialType;
-import phanes.replay.user.dto.SocialProfile;
+import phanes.replay.user.dto.oauth.SocialProfile;
 
 public interface SocialLoginStrategy {
     String getSocialUrl(String state);
